Read os-release instead of using platform.linux_distribution()
The function was removed in https://bugs.python.org/issue1322
and cannot be used in python3.8.

There are replacements outside of the stdlib, but it doesn't seem
worth it it add a dependency on another module. Instead, a simple
parser for os-release is implemented. os-release is present on all
distros from the last few years and PRETTY_NAME gives a reliable
display name of the distro without any heuristics.