Fix exception from accessing cpu_info when it is None
If `cpuinfo` fails to import, then `cpu_info` is set to `None` at the top of this file.  But in the `sse2` check, `cpu_info` is accessed without first checking if it is `None`, causing the build to fail.

This fix errs on the conservative side, assuming that we do not want to build SSE2 if `cpu_info` could not be imported.  It's possible a similar change is needed on the next line for AVX2, but I wasn't sure why there was no `'avx2' not in cpu_info['flags']` condition there.